createSccess & updateSuccess hooked to notifications + Flash styles
