+++ src/components/Flash/index.css
@@ -1,0 +1,6 @@
+.Flash {
+  position: fixed;
+  left: 20%;
+  width: 60%;
+  bottom: 5%;
+}
